requirements.txt
Flask>=2.0.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
albumentations>=1.3.0
opencv-python>=4.5.0
Pillow>=9.0.0
numpy>=1.20.0
git+https://github.com/openai/CLIP.git